Refactor cursor classes to separate package
