+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/cursor/Cursor.java
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.executor.cursor;
+package com.amazon.opendistroforelasticsearch.sql.cursor;
 
 
 public interface Cursor {
