+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/cursor/CursorType.java
@@ -13,26 +13,36 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.executor.cursor;
+package com.amazon.opendistroforelasticsearch.sql.cursor;
 
-/**
- * A placeholder Cursor implementation to work with non-paginated queries.
- */
-public class NullCursor implements Cursor {
+import java.util.HashMap;
+import java.util.Map;
 
-    private final CursorType type = CursorType.NULL;
+public enum CursorType {
+    NULL(null),
+    DEFAULT("d"),
+    AGGREGATION("a"),
+    JOIN("j");
 
-    @Override
-    public String generateCursorId() {
-        return null;
+    public String id;
+
+    CursorType(String id) {
+        this.id = id;
     }
 
-    @Override
-    public CursorType getType() {
-        return type;
+    public String getId() {
+        return this.id;
     }
 
-    public NullCursor from(String cursorId) {
-        return new NullCursor();
+    public static final Map<String, CursorType> LOOKUP = new HashMap<>();
+
+    static {
+        for (CursorType type : CursorType.values()) {
+            LOOKUP.put(type.getId(), type);
+        }
+    }
+
+    public static CursorType getById(String id) {
+        return LOOKUP.getOrDefault(id, NULL);
     }
 }
