+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/executor/format/PrettyFormatRestExecutor.java
@@ -18,8 +18,8 @@ package com.amazon.opendistroforelasticsearch.sql.executor.format;
 import com.amazon.opendistroforelasticsearch.sql.exception.SqlParseException;
 import com.amazon.opendistroforelasticsearch.sql.executor.QueryActionElasticExecutor;
 import com.amazon.opendistroforelasticsearch.sql.executor.RestExecutor;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.DefaultCursor;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.NullCursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.DefaultCursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.NullCursor;
 import com.amazon.opendistroforelasticsearch.sql.query.DefaultQueryAction;
 import com.amazon.opendistroforelasticsearch.sql.query.QueryAction;
 import com.amazon.opendistroforelasticsearch.sql.query.join.BackOffRetryStrategy;
