+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/executor/format/Protocol.java
@@ -20,8 +20,8 @@ import com.amazon.opendistroforelasticsearch.sql.domain.Delete;
 import com.amazon.opendistroforelasticsearch.sql.domain.IndexStatement;
 import com.amazon.opendistroforelasticsearch.sql.domain.Query;
 import com.amazon.opendistroforelasticsearch.sql.domain.QueryStatement;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.Cursor;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.NullCursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.Cursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.NullCursor;
 import com.amazon.opendistroforelasticsearch.sql.executor.format.DataRows.Row;
 import com.amazon.opendistroforelasticsearch.sql.executor.format.Schema.Column;
 import com.amazon.opendistroforelasticsearch.sql.executor.adapter.QueryPlanQueryAction;
