+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/executor/format/SelectResultSet.java
@@ -29,8 +29,8 @@ import com.amazon.opendistroforelasticsearch.sql.domain.TableOnJoinSelect;
 import com.amazon.opendistroforelasticsearch.sql.esdomain.mapping.FieldMapping;
 import com.amazon.opendistroforelasticsearch.sql.exception.SqlFeatureNotImplementedException;
 import com.amazon.opendistroforelasticsearch.sql.executor.Format;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.Cursor;
-import com.amazon.opendistroforelasticsearch.sql.executor.cursor.DefaultCursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.Cursor;
+import com.amazon.opendistroforelasticsearch.sql.cursor.DefaultCursor;
 import com.amazon.opendistroforelasticsearch.sql.metrics.MetricName;
 import com.amazon.opendistroforelasticsearch.sql.metrics.Metrics;
 import com.amazon.opendistroforelasticsearch.sql.utils.SQLFunctions;
